+++ src/frontend/apps/e2e/__tests__/app-impress/doc-editor.spec.ts
@@ -5,7 +5,6 @@ import cs from 'convert-stream';
 
 import {
   createDoc,
-  getMenuItem,
   goToGridDoc,
   overrideConfig,
   verifyDocName,
@@ -148,7 +147,7 @@ test.describe('Doc Editor', () => {
     const wsClosePromise = webSocket.waitForEvent('close');
 
     await selectVisibility.click();
-    await getMenuItem(page, 'Connected').click();
+    await page.getByRole('menuitemradio', { name: 'Connected' }).click();
 
     // Assert that the doc reconnects to the ws
     const wsClose = await wsClosePromise;
@@ -605,7 +604,7 @@ test.describe('Doc Editor', () => {
     await page.getByRole('button', { name: 'Share' }).click();
 
     await page.getByTestId('doc-access-mode').click();
-    await getMenuItem(page, 'Reading').click();
+    await page.getByRole('menuitemradio', { name: 'Reading' }).click();
 
     // Close the modal
     await page.getByRole('button', { name: 'close' }).first().click();