+++ src/frontend/apps/e2e/__tests__/app-impress/doc-grid-move.spec.ts
@@ -3,7 +3,6 @@ import { expect, test } from '@playwright/test';
 import {
   createDoc,
   getGridRow,
-  getMenuItem,
   getOtherBrowserName,
   mockedListDocs,
   toggleHeaderMenu,
@@ -207,7 +206,7 @@ test.describe('Doc grid move', () => {
     const row = await getGridRow(page, titleDoc1);
     await row.getByText(`more_horiz`).click();
 
-    await getMenuItem(page, 'Move into a doc').click();
+    await page.getByRole('menuitem', { name: 'Move into a doc' }).click();
 
     await expect(
       page.getByRole('dialog').getByRole('heading', { name: 'Move' }),
@@ -295,7 +294,7 @@ test.describe('Doc grid move', () => {
     const row = await getGridRow(page, titleDoc1);
     await row.getByText(`more_horiz`).click();
 
-    await getMenuItem(page, 'Move into a doc').click();
+    await page.getByRole('menuitem', { name: 'Move into a doc' }).click();
 
     await expect(
       page.getByRole('dialog').getByRole('heading', { name: 'Move' }),
@@ -342,7 +341,9 @@ test.describe('Doc grid move', () => {
       `doc-share-access-request-row-${emailRequest}`,
     );
     await container.getByTestId('doc-role-dropdown').click();
-    await getMenuItem(otherPage, 'Administrator').click();
+    await otherPage
+      .getByRole('menuitemradio', { name: 'Administrator' })
+      .click();
     await container.getByRole('button', { name: 'Approve' }).click();
 
     await expect(otherPage.getByText('Access Requests')).toBeHidden();
@@ -353,7 +354,7 @@ test.describe('Doc grid move', () => {
     await page.reload();
     await row.getByText(`more_horiz`).click();
 
-    await getMenuItem(page, 'Move into a doc').click();
+    await page.getByRole('menuitem', { name: 'Move into a doc' }).click();
 
     await expect(
       page.getByRole('dialog').getByRole('heading', { name: 'Move' }),