+++ src/frontend/apps/e2e/__tests__/app-impress/doc-member-create.spec.ts
@@ -3,7 +3,6 @@ import { expect, test } from '@playwright/test';
 import {
   BROWSERS,
   createDoc,
-  getMenuItem,
   keyCloakSignIn,
   randomName,
   verifyDocName,
@@ -111,13 +110,21 @@ test.describe('Document create member', () => {
 
     // Check roles are displayed
     await list.getByTestId('doc-role-dropdown').click();
-    await expect(getMenuItem(page, 'Reader')).toBeVisible();
-    await expect(getMenuItem(page, 'Editor')).toBeVisible();
-    await expect(getMenuItem(page, 'Owner')).toBeVisible();
-    await expect(getMenuItem(page, 'Administrator')).toBeVisible();
+    await expect(
+      page.getByRole('menuitemradio', { name: 'Reader' }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitemradio', { name: 'Editor' }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitemradio', { name: 'Owner' }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitemradio', { name: 'Administrator' }),
+    ).toBeVisible();
 
     // Validate
-    await getMenuItem(page, 'Administrator').click();
+    await page.getByRole('menuitemradio', { name: 'Administrator' }).click();
     await page.getByTestId('doc-share-invite-button').click();
 
     // Check invitation added
@@ -163,7 +170,7 @@ test.describe('Document create member', () => {
     // Choose a role
     const container = page.getByTestId('doc-share-add-member-list');
     await container.getByTestId('doc-role-dropdown').click();
-    await getMenuItem(page, 'Owner').click();
+    await page.getByRole('menuitemradio', { name: 'Owner' }).click();
 
     const responsePromiseCreateInvitation = page.waitForResponse(
       (response) =>
@@ -181,7 +188,7 @@ test.describe('Document create member', () => {
 
     // Choose a role
     await container.getByTestId('doc-role-dropdown').click();
-    await getMenuItem(page, 'Owner').click();
+    await page.getByRole('menuitemradio', { name: 'Owner' }).click();
 
     const responsePromiseCreateInvitationFail = page.waitForResponse(
       (response) =>
@@ -218,7 +225,7 @@ test.describe('Document create member', () => {
     // Choose a role
     const container = page.getByTestId('doc-share-add-member-list');
     await container.getByTestId('doc-role-dropdown').click();
-    await getMenuItem(page, 'Administrator').click();
+    await page.getByRole('menuitemradio', { name: 'Administrator' }).click();
 
     const responsePromiseCreateInvitation = page.waitForResponse(
       (response) =>
@@ -245,13 +252,13 @@ test.describe('Document create member', () => {
     );
 
     await userInvitation.getByTestId('doc-role-dropdown').click();
-    await getMenuItem(page, 'Reader').click();
+    await page.getByRole('menuitemradio', { name: 'Reader' }).click();
 
     const responsePatchInvitation = await responsePromisePatchInvitation;
     expect(responsePatchInvitation.ok()).toBeTruthy();
 
     await userInvitation.getByTestId('doc-role-dropdown').click();
-    await getMenuItem(page, 'Remove access').click();
+    await page.getByRole('menuitemradio', { name: 'Remove access' }).click();
 
     await expect(userInvitation).toBeHidden();
   });
@@ -303,7 +310,7 @@ test.describe('Document create member', () => {
       `doc-share-access-request-row-${emailRequest}`,
     );
     await container.getByTestId('doc-role-dropdown').click();
-    await getMenuItem(page, 'Administrator').click();
+    await page.getByRole('menuitemradio', { name: 'Administrator' }).click();
     await container.getByRole('button', { name: 'Approve' }).click();
 
     await expect(page.getByText('Access Requests')).toBeHidden();