+++ src/frontend/apps/e2e/__tests__/app-impress/utils-share.ts
@@ -2,7 +2,6 @@ import { Page, chromium, expect } from '@playwright/test';
 
 import {
   BrowserName,
-  getMenuItem,
   getOtherBrowserName,
   keyCloakSignIn,
   verifyDocName,
@@ -40,7 +39,7 @@ export const addNewMember = async (
 
   // Choose a role
   await page.getByTestId('doc-role-dropdown').click();
-  await getMenuItem(page, role).click();
+  await page.getByRole('menuitemradio', { name: role }).click();
   await page.getByTestId('doc-share-invite-button').click();
 
   return users[index].email;
@@ -52,7 +51,7 @@ export const updateShareLink = async (
   linkRole?: LinkRole | null,
 ) => {
   await page.getByTestId('doc-visibility').click();
-  await getMenuItem(page, linkReach).click();
+  await page.getByRole('menuitemradio', { name: linkReach }).click();
 
   const visibilityUpdatedText = page
     .getByText('The document visibility has been updated')
@@ -62,7 +61,7 @@ export const updateShareLink = async (
 
   if (linkRole) {
     await page.getByTestId('doc-access-mode').click();
-    await getMenuItem(page, linkRole).click();
+    await page.getByRole('menuitemradio', { name: linkRole }).click();
     await expect(visibilityUpdatedText).toBeVisible();
   }
 };
@@ -77,7 +76,7 @@ export const updateRoleUser = async (
   const currentUser = list.getByTestId(`doc-share-member-row-${email}`);
   const currentUserRole = currentUser.getByTestId('doc-role-dropdown');
   await currentUserRole.click();
-  await getMenuItem(page, role).click();
+  await page.getByRole('menuitemradio', { name: role }).click();
   await list.click();
 };